fix: 강의 조회 시 신청률 높은 순으로 가능하도록 수정

+++ src/main/java/com/weolbu/LMS/dtos/CourseResponse.java
@@ -20,5 +20,5 @@ public class CourseResponse {
     private Long price;
     private LocalDateTime createdDateTime;
     private Long registrationCount;
-    private Long registrationRate;
+    private Double registrationRate;
 }
